Shimless: Update mojo int type to match proto definition
The integer type for the index variables was updated in RMAD service to
reflect potential negative values.

Bug: 1198187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3403260
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Cr-Commit-Position: refs/heads/main@{#961622}

+++ ash/webui/shimless_rma/backend/shimless_rma_service.cc
@@ -604,9 +604,9 @@ void ShimlessRmaService::GetOriginalDramPartNumber(
 
 void ShimlessRmaService::SetDeviceInformation(
     const std::string& serial_number,
-    uint8_t region_index,
-    uint8_t sku_index,
-    uint8_t white_label_index,
+    int32_t region_index,
+    int32_t sku_index,
+    int32_t white_label_index,
     const std::string& dram_part_number,
     SetDeviceInformationCallback callback) {
   if (state_proto_.state_case() != rmad::RmadState::kUpdateDeviceInfo) {
